backend/create-booking-v2/requirements.txt
psycopg2-binary==2.9.9
requests>=2.28.0